fixed issue #3 and unknown issue
in addition to needing better handling for low balance emails, there was a new issue that was causing the script to fail every time.  i did not realize the timed trigger was passing an object into the chosen function.  and the function was not set to handle the object.  it was expecting string or undefined.